refactor(controller): InputValueConverter의 클래스명과 메서드명을 무슨 일을 하는지 명확히 표현하도록 변경
